Restructure and clean up
